backend/requirements.txt
fastapi
uvicorn
duckdb
anthropic
pandas
reportlab
python-multipart
python-dotenv

